pyproject.toml
[tool.poetry]
name = "vnasrbench"
version = "0.1.0"
description = "Vietnamese Streaming ASR Benchmark: offline + streaming simulation, reproducible evaluation"
authors = ["Your Name <you@email>"]
readme = "README.md"
packages = [{ include = "vnasrbench", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
numpy = "^1.26.4"
torch = {version = "^2.3.1", optional = true}
torchaudio = {version = "^2.3.1", optional = true}
transformers = "^4.44.2"
accelerate = "^0.34.2"
soundfile = "^0.12.1"
jiwer = "^3.0.4"
pyyaml = "^6.0.2"
tqdm = "^4.66.5"
rich = "^13.8.1"
pandas = "^2.2.2"
matplotlib = "^3.9.2"

[tool.poetry.extras]
gpu = ["torch", "torchaudio"]

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
ruff = "^0.6.4"
mypy = "^1.11.2"
pre-commit = "^3.8.0"

[tool.poetry.scripts]
vnasrbench = "vnasrbench.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
